+++ projects/GKCore/GKCore/Kinships/KinshipsCulture.cs
@@ -18,13 +18,14 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+using System;
 using System.Collections.Generic;
 
 namespace GKCore.Kinships
 {
     internal sealed class KSTypeEntry
     {
-        public int Id;
+        public int Index;
         public string Sign;
         public string Name;
         public bool HasExt;
@@ -36,21 +37,21 @@ namespace GKCore.Kinships
 
         public override string ToString()
         {
-            return $"{Sign} / {Name} / {HasExt} / {Description}";
+            return $"{Index} / {Sign} / {Name} / {HasExt} / {Description}";
         }
 
         public void Prepare()
         {
-            if (Name.StartsWith("@")) {
-                int lsid = int.Parse(Name.Substring(1));
-                Name = LangMan.LS((LSID)lsid);
-            }
+            Name = KinshipsCulture.GetLocalizedString(Name);
         }
     }
 
 
     internal sealed class KSDefinitionEntry
     {
+        public int Index;
+        public bool Enable;
+
         public string[] PrevPrevious;
         public string[] Previous;
         public string[] Current;
@@ -58,7 +59,6 @@ namespace GKCore.Kinships
 
         public short Great;
         public short Degree;
-
 
         public HashSet<int> PrevPrevRels;
         public HashSet<int> PrevRels;
@@ -76,7 +76,7 @@ namespace GKCore.Kinships
 
         public override string ToString()
         {
-            return $"[{string.Join(",", PrevPrevious)}] / [{string.Join(",", Previous)}] / [{string.Join(",", Current)}] / {Result} / {Great} / {Degree}";
+            return $"{Index} / {Enable} / [{string.Join(",", PrevPrevious)}] / [{string.Join(",", Previous)}] / [{string.Join(",", Current)}] / {Result} / {Great} / {Degree}";
         }
 
         public void Prepare(Dictionary<string, int> indexes)
@@ -85,20 +85,20 @@ namespace GKCore.Kinships
             ProcessTypes(indexes, Previous, PrevRels);
             ProcessTypes(indexes, Current, CurrRels);
 
-            if (indexes.TryGetValue(Result, out int resId)) {
-                FinRel = resId;
+            if (indexes.TryGetValue(Result, out int idx)) {
+                FinRel = idx;
             } else {
-                // index to log
+                Logger.WriteError($"KSDefinitionEntry.Prepare({Index}): Type signature `{Result}` does not match index");
             }
         }
 
-        private static void ProcessTypes(Dictionary<string, int> indexes, string[] array, HashSet<int> hashSet)
+        private void ProcessTypes(Dictionary<string, int> indexes, string[] array, HashSet<int> hashSet)
         {
             foreach (var str in array) {
                 if (indexes.TryGetValue(str, out int idx)) {
                     hashSet.Add(idx);
                 } else {
-                    // index to log
+                    Logger.WriteError($"KSDefinitionEntry.Prepare({Index}): Type signature `{str}` does not match index");
                 }
             }
         }
@@ -110,6 +110,8 @@ namespace GKCore.Kinships
     /// </summary>
     internal class KinshipsCulture
     {
+        public string GreatPrefix { get; set; }
+
         public KSTypeEntry[] Types { get; set; }
         public KSDefinitionEntry[] Definitions { get; set; }
 
@@ -120,14 +122,44 @@ namespace GKCore.Kinships
         {
             Types = new KSTypeEntry[0];
             Definitions = new KSDefinitionEntry[0];
-
-            KinshipTypeIndexes = new int[(int)KinshipType.ktLast + 1];
         }
 
         public KSTypeEntry GetTypeEntry(int id)
         {
             int index = KinshipTypeIndexes[id];
             return Types[index];
+        }
+
+        public void Prepare()
+        {
+            GreatPrefix = GetLocalizedString(GreatPrefix);
+
+            KinshipTypeIndexes = new int[(int)KinshipType.ktLast + 1];
+            for (KinshipType kt = KinshipType.ktNone; kt < KinshipType.ktLast; kt++) {
+                var sign = kt.ToString();
+                var index = Array.FindIndex(Types, (kte) => kte.Sign == sign);
+                var ktEntry = Types[index];
+                KinshipTypeIndexes[(int)kt] = index;
+            }
+
+            var indexes = new Dictionary<string, int>();
+            foreach (var kinType in Types) {
+                indexes.Add(kinType.Sign, kinType.Index);
+                kinType.Prepare();
+            }
+
+            foreach (var kinDef in Definitions) {
+                kinDef.Prepare(indexes);
+            }
+        }
+
+        public static string GetLocalizedString(string value)
+        {
+            if (value.StartsWith("@")) {
+                int lsid = int.Parse(value.Substring(1));
+                value = LangMan.LS((LSID)lsid);
+            }
+            return value;
         }
     }
 }
